+++ CHANGELOG.md
@@ -52,6 +52,7 @@
 - `ContentRepresentation`
 - `Elements.Door`
 - `ComponentBase.UseRepresentationInstances` - an option flag to make generating fitting models faster/smaller.
+- `ContentConfiguration.AllowRotatation`
 
 ### Fixed